common/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bidict==0.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
facenet-pytorch==2.6.0
fastapi==0.115.5
fer==22.5.1
ffmpeg==1.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
h11==0.14.0
h5py==3.12.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
keras==3.7.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
mediapipe==0.10.18
ml_dtypes==0.5.0
moviepy==2.1.1
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
proglog==0.1.10
protobuf==4.25.5
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
requests==2.32.3
rich==13.9.4
scipy==1.13.1
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
socketIO-client==0.7.2
sounddevice==0.5.1
starlette==0.41.3
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
zipp==3.21.0
